Fix the Feature Launcher test reliability
Some of the Feature Launcher tests deploy the Gogo Shell. If the Gogo shell detects that stdin has closed then it shuts down the framework automatically. When running in CI (and elsewhere) we use batch mode with no terminal which disconnects the test run and gives it an empty stdin. This causes the tests to fail as gogo shuts things down before the test finishes. Instead we must configure gogo to be non-interactive during the tests.

Signed-off-by: Tim Ward <timothyjward@apache.org>