Fix UnsupportedOperationException from OperationBuilders
When registering artifact repositories with an Operation Builder the user gets an UnsupportedOperationException if the builder is using default repositories. Furthermore, setting useDefaultRepositories causes any previously registered repositories to be lost. This should be fixed to ensure that no registered repositories are lost, and that default repositories are only used if enabled and not overriden in the builder.

Signed-off-by: Tim Ward <timothyjward@apache.org>